requirements.txt
asgiref==3.7.2
certifi==2024.2.2
charset-normalizer==3.3.2
dj-database-url==2.1.0
Django==5.0.2
django-cors-headers==4.3.1
django-environ==0.11.2
djangorestframework==3.14.0
environ==1.0
gunicorn==21.2.0
hubspot==0.1.14.dev0
hubspot-api-client==8.2.1
idna==3.6
packaging==23.2
python-dateutil==2.8.2
python-decouple==3.8
pytz==2024.1
ramapi==0.1.0
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.9.0
urllib3==2.2.1
whitenoise==6.6.0
